+++ test/bigtest/network/config.js
@@ -544,15 +544,15 @@ export default function configure() {
     totalRecords: 0
   });
 
-  this.get('/source-storage/formattedRecords', {
+  this.get('/source-storage/records', {
     records: [],
     totalRecords: 0
   });
-  this.get('/source-storage/formattedRecords/:id', {});
+  this.get('/source-storage/records/:id/formatted', {});
 
   this.get('/inventory/config/instances/blocked-fields');
 
-  this.get('/source-storage/formattedRecords/:id', {
+  this.get('/source-storage/records/:id/formatted', {
     instanceTypes: [],
     totalRecords: 0
   });