+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Select and move holdings with items or items to another instance. Refs UIIN-1098.
 * Move items between holdings. Refs UIIN-1097.
 * Disable fields in Preceding/Succeeding section, when Instance is controlled by Source MARC. Refs UIIN-1128.
+* Update Requests to SRS for v4. Refs UIIN-1158.
 * Show two instance records as a split screen. Refs UIIN-1103.
 
 ## [3.0.1] (IN PROGRESS)